stock: completing

+++ stocks/internal/usecase/model.go
@@ -4,7 +4,7 @@ import "stocks/internal/models"
 
 type GetSKU struct {
 	Name   string
-	Price  float64
+	Price  uint32
 	Count  int
 	Type   string
 	UserId models.UserID
